Camera, parallax and looping texture: smooth movement + parallax widescreen (#356)
Replacing a lot of positioning variable from int to float makes a lot of scenes much smoother, as the positioning is stored in psx pixel values. Includes:

Camera movements
Parallax
Scrolling textures
Also, a lot of screens were not possible to make widescreen due to the parallax background not covering the entire screen (branbal), they are now custom extended when displayed, which now covers the entire 16:9 screen. This concerns >8 backgrounds. #335

+fix for clayra trunk text not showing #367
+fix for previous regression where battle aspect was elongated #364
+a lot of widescreen fixes for specific maps